Move checking attachments to its own helper function
