+++ src/bot.ts
@@ -8,6 +8,7 @@ import { EmbedBuilder } from 'discord.js';
 
 import type { Command, MessageData, ScamWSData, serverDBData } from './types';
 import jsQR from 'jsqr';
+import { checkAttachments } from './helpers/checkers';
 
 process.on("message", (msg: MessageData) => {
   if (!msg.type) return false;
@@ -215,21 +216,7 @@ bot.on('messageCreate', async message => {
     }
   }
 
-  message.attachments.forEach(async (attachment) => {
-    if (!attachment.contentType?.includes('image')) return;
-    const image = await Jimp.read(attachment.url);
-    const code = jsQR(image.bitmap.data as any, image.bitmap.width, image.bitmap.height);
-    if (code === null) return;
-    if (code.data.startsWith('https://discord.com/ra/') || code.data.startsWith('https://discordapp.com/ra/')) {
-      message.reply({
-        'embeds': [{
-          'description': ':warning: POSSIBLE SCAM DETECTED :warning:\n\nThe image above contains a Discord Login QR code.\nScanning this code with the Discord app will give whoever made the code FULL ACCESS to your account',
-        }]
-      }).catch((err) => {
-        if (err && message.deletable) message.delete();
-      });
-    }
-  });
+  await checkAttachments(message);
 
   // Check if command exists
   const command = commands.find(command => command.data.name === cmd);
